✨ (use-post-add-flow.ts): Add useGetRefreshFlows hook to refresh flows after adding a new flow ♻️ (use-add-flow.ts): Remove unnecessary refreshFlows hook and refactor code to use useGetRefreshFlows hook instead 📝 (dropdown/index.tsx): Update import paths for useDuplicateFlows and useSelectOptionsChange hooks 📝 (grid/index.tsx): Update import path for useDescriptionModal hook and remove commented out code related to playground functionality
✨ (MainPage): Add new hooks for handling duplicate flows, selecting all flows, changing select options, and managing selected flows
📝 (MainPage): Update import path for useDescriptionModal hook to correct location
🔧 (MainPage): Remove unused code related to playground functionality in ListComponent

🔧 (use-handle-duplicate.tsx): Remove unused file use-handle-duplicate.tsx
🔧 (componentsComponent/index.tsx): Remove unused imports and function call for useDuplicateFlows
📝 (homePage/index.tsx): Add console log to log data.flows for debugging purposes
