+++ src/frontend/src/hooks/flows/use-add-flow.ts
@@ -36,7 +36,6 @@ const useAddFlow = () => {
   const myCollectionId = useFolderStore((state) => state.myCollectionId);
 
   const { mutate: postAddFlow } = usePostAddFlow();
-  const { mutate: refreshFlows } = useGetRefreshFlows();
 
   const addFlow = async (params?: {
     flow?: FlowType;
@@ -57,8 +56,6 @@ const useAddFlow = () => {
         );
       });
       // Create a new flow with a default name if no flow is provided.
-      const folder_id = folderId ?? myCollectionId ?? "";
-
       if (params?.override && flow) {
         const flowId = flows?.find((f) => f.name === flow.name);
         if (flowId) {
@@ -66,12 +63,11 @@ const useAddFlow = () => {
         }
       }
 
+      const folder_id = folderId ?? myCollectionId ?? "";
       const flowsToCheckNames = flows?.filter(
         (f) => f.folder_id === myCollectionId,
       );
-
       const newFlow = createNewFlow(flowData!, folder_id, flow);
-
       const newName = addVersionToDuplicates(newFlow, flowsToCheckNames ?? []);
       newFlow.name = newName;
       newFlow.folder_id = folder_id;
@@ -91,11 +87,6 @@ const useAddFlow = () => {
               ["saved_components"]: data,
             }),
           }));
-
-          refreshFlows({
-            get_all: true,
-            header_flows: true,
-          });
 
           setFlowToCanvas(createdFlow);
           resolve(createdFlow.id);
