+++ src/frontend/src/pages/MainPage/components/dropdown/index.tsx
@@ -3,8 +3,8 @@ import { DropdownMenuItem } from "@/components/ui/dropdown-menu";
 import useAlertStore from "@/stores/alertStore";
 import { FlowType } from "@/types/flow";
 import { downloadFlow } from "@/utils/reactflowUtils";
-import useDuplicateFlows from "../../oldComponents/componentsComponent/hooks/use-handle-duplicate";
-import useSelectOptionsChange from "../../oldComponents/componentsComponent/hooks/use-select-options-change";
+import useDuplicateFlows from "../../hooks/use-handle-duplicate";
+import useSelectOptionsChange from "../../hooks/use-select-options-change";
 
 type DropdownComponentProps = {
   flowData: FlowType;
@@ -15,7 +15,6 @@ type DropdownComponentProps = {
 const DropdownComponent = ({
   flowData,
   setOpenDelete,
-  handlePlaygroundClick,
 }: DropdownComponentProps) => {
   const setSuccessData = useAlertStore((state) => state.setSuccessData);
   const setErrorData = useAlertStore((state) => state.setErrorData);
@@ -23,11 +22,7 @@ const DropdownComponent = ({
   const { handleDuplicate } = useDuplicateFlows(
     [flowData.id],
     [flowData],
-    () => {},
     setSuccessData,
-    () => {},
-    () => {},
-    "flow",
   );
 
   const handleExport = () => {
@@ -45,30 +40,6 @@ const DropdownComponent = ({
 
   return (
     <>
-      {/* <DropdownMenuItem onClick={() => {}} className="cursor-pointer">
-        <ForwardedIconComponent
-          name="square-pen"
-          aria-hidden="true"
-          className="mr-2 h-4 w-4"
-        />
-        Edit details
-      </DropdownMenuItem> */}
-      {/* {handlePlaygroundClick && (
-        <DropdownMenuItem
-          onClick={(e) => {
-            e.stopPropagation();
-            handlePlaygroundClick();
-          }}
-          className="cursor-pointer sm:hidden"
-        >
-          <ForwardedIconComponent
-            name="play"
-            aria-hidden="true"
-            className="mr-2 h-4 w-4"
-          />
-          Playground
-        </DropdownMenuItem>
-      )} */}
       <DropdownMenuItem
         onClick={(e) => {
           e.stopPropagation();
