+++ src/frontend/src/pages/MainPage/components/list/index.tsx
@@ -17,20 +17,18 @@ import { swatchColors } from "@/utils/styleUtils";
 import { cn, getNumberFromString } from "@/utils/utils";
 import { useState } from "react";
 import { useParams } from "react-router-dom";
-import useDescriptionModal from "../../oldComponents/componentsComponent/hooks/use-description-modal";
+import useDescriptionModal from "../../hooks/use-description-modal";
 import { useGetTemplateStyle } from "../../utils/get-template-style";
 import { timeElapsed } from "../../utils/time-elapse";
 import DropdownComponent from "../dropdown";
 
 const ListComponent = ({ flowData }: { flowData: FlowType }) => {
   const navigate = useCustomNavigate();
-  /* const [openPlayground, setOpenPlayground] = useState(false);
-  const [loadingPlayground, setLoadingPlayground] = useState(false); */
+
   const [openDelete, setOpenDelete] = useState(false);
   const setSuccessData = useAlertStore((state) => state.setSuccessData);
   const { deleteFlow } = useDeleteFlow();
   const setErrorData = useAlertStore((state) => state.setErrorData);
-  /* const setCurrentFlow = useFlowsManagerStore((state) => state.setCurrentFlow); */
   const { folderId } = useParams();
   const isComponent = flowData.is_component ?? false;
   const setFlowToCanvas = useFlowsManagerStore(
@@ -39,38 +37,6 @@ const ListComponent = ({ flowData }: { flowData: FlowType }) => {
   const { getIcon } = useGetTemplateStyle(flowData);
 
   const editFlowLink = `/flow/${flowData.id}${folderId ? `/folder/${folderId}` : ""}`;
-
-  /* function hasPlayground(flow?: FlowType) {
-    if (!flow) {
-      return false;
-    }
-    const { inputs, outputs } = getInputsAndOutputs(flow?.data?.nodes ?? []);
-    return inputs.length > 0 || outputs.length > 0;
-  } */
-
-  /* const handlePlaygroundClick = () => {
-    track("Playground Button Clicked", { flowId: flowData.id });
-    setLoadingPlayground(true);
-
-    if (flowData) {
-      if (!hasPlayground(flowData)) {
-        setErrorData({
-          title: "Error",
-          list: ["This flow doesn't have a playground."],
-        });
-        setLoadingPlayground(false);
-        return;
-      }
-      setCurrentFlow(flowData);
-      setOpenPlayground(true);
-      setLoadingPlayground(false);
-    } else {
-      setErrorData({
-        title: "Error",
-        list: ["Error getting flow data."],
-      });
-    }
-  }; */
 
   const handleClick = async () => {
     if (!isComponent) {
@@ -155,22 +121,6 @@ const ListComponent = ({ flowData }: { flowData: FlowType }) => {
 
         {/* right side */}
         <div className="ml-5 flex items-center gap-2">
-          {/* {flowData.is_component ? (
-            <></>
-          ) : (
-            <Button
-              variant="outline"
-              disabled={loadingPlayground || !hasPlayground(flowData)}
-              onClick={(e) => {
-                e.preventDefault();
-                e.stopPropagation();
-                handlePlaygroundClick();
-              }}
-              className="hidden sm:block"
-            >
-              Playground
-            </Button>
-          )} */}
           <DropdownMenu>
             <DropdownMenuTrigger asChild>
               <Button
@@ -202,16 +152,7 @@ const ListComponent = ({ flowData }: { flowData: FlowType }) => {
           </DropdownMenu>
         </div>
       </Card>
-      {/* {openPlayground && (
-        <IOModal
-          key={flowData.id}
-          cleanOnClose={true}
-          open={openPlayground}
-          setOpen={setOpenPlayground}
-        >
-          <></>
-        </IOModal>
-      )} */}
+
       {openDelete && (
         <DeleteConfirmationModal
           open={openDelete}
